commoit

+++ workout tracker/src/main/java/workout/repo/CategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends CrudRepository<WorkoutCategory,Integer> {
+public interface CategoryRepository extends CrudRepository<WorkoutCategory,Long> {
 }
